Move ListQuery.ts to list folder
