+++ packages/client/src/observable/internal/list/ListsHelper.ts
@@ -19,12 +19,12 @@ import type { ListPayload } from "../../ListPayload.js";
 import type { ObserveListOptions } from "../../ObservableClient.js";
 import type { Observer } from "../../ObservableClient/common.js";
 import { AbstractHelper } from "../AbstractHelper.js";
-import { ListQuery } from "../ListQuery.js";
 import type { OrderByCanonicalizer } from "../OrderByCanonicalizer.js";
 import type { QuerySubscription } from "../QuerySubscription.js";
 import type { Store } from "../Store.js";
 import type { WhereClauseCanonicalizer } from "../WhereClauseCanonicalizer.js";
 import type { ListCacheKey } from "./ListCacheKey.js";
+import { ListQuery } from "./ListQuery.js";
 
 export class ListsHelper extends AbstractHelper<
   ListQuery,
